+++ src/components/cashBalances/cashBalances.utils.js
@@ -1,0 +1,6 @@
+export function getBillsTotal(bills) {
+  return bills.reduce(
+    (sum, b) => sum + (Number(b.denomination) * Number(b.quantity)),
+    0
+  )
+}