+++ src/components/expenses/ExpensesPage/index.jsx
@@ -1,11 +1,11 @@
 import { DropdownControlled, TextControlled } from '@/common/components/form';
 import { Filters, Table } from '@/common/components/table';
-import { ENTITIES, PAGES } from "@/common/constants";
+import { ENTITIES, PAGES, SELECT_ALL_OPTION } from "@/common/constants";
 import { createFilter } from '@/common/utils';
 import { useFilters } from "@/hooks";
 import { FormProvider } from "react-hook-form";
 import { Form } from "semantic-ui-react";
-import { EMPTY_FILTERS, EXPENSES_FILTERS_KEY, EXPENSES_STATE_OPTIONS, HEADERS } from '../expenses.constants';
+import { EMPTY_FILTERS, EXPENSES_FILTERS_KEY, EXPENSES_STATE_OPTIONS, EXPENSE_STATES, HEADERS } from '../expenses.constants';
 
 const ExpensesPage = ({ expenses = [], isLoading, onRefetch, onDownloadExcel }) => {
 
@@ -17,11 +17,9 @@ const ExpensesPage = ({ expenses = [], isLoading, onRefetch, onDownloadExcel }) 
     methods
   } = useFilters({ defaultFilters: EMPTY_FILTERS, key: EXPENSES_FILTERS_KEY });
 
-  const onFilter = createFilter(filters, [
-    "id",
-    "name",
-    { field: "categories", nestedField: "name" }
-  ]);
+  const onFilter = createFilter(filters, ['id', 'name', { field: "categories", nestedField: "name" }], {
+    allState: SELECT_ALL_OPTION.value,
+  });
 
   return (
     <>
@@ -36,7 +34,7 @@ const ExpensesPage = ({ expenses = [], isLoading, onRefetch, onDownloadExcel }) 
               width="200px"
               name="state"
               options={EXPENSES_STATE_OPTIONS}
-              value={EMPTY_FILTERS.state}
+              value={SELECT_ALL_OPTION.state}
               afterChange={onSubmit}
             />
             <TextControlled name="id" placeholder="Id" width="100px" />
@@ -55,6 +53,7 @@ const ExpensesPage = ({ expenses = [], isLoading, onRefetch, onDownloadExcel }) 
         filters={filters}
         setFilters={setFilters}
         onDownloadExcel={onDownloadExcel}
+        color={EXPENSE_STATES[filters.state]?.color}
       />
     </>
   );